+++ Calculations/Weights.lua
@@ -1,0 +1,1 @@
+-- Component weighting and boss fight calculations